fastapi_legacy/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
email-validator
httpx
python-multipart
sqladmin
greenlet
pytest
pytest-asyncio
